backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
pypdf==3.17.0
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.2
pandas==2.1.3

# ML and embeddings
# Upgrade to avoid deprecated cached_download usage
sentence-transformers==2.6.1
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.2.1
# Align with newer sentence-transformers & transformers
huggingface-hub==0.34.0

# Google Gemini
google-generativeai==0.3.0

# CORS and async
aiofiles==23.2.1
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1